+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/dto/OrderDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.UUID;
 
@@ -22,7 +23,7 @@ public class OrderDto {
     private double deliveryWeight;
     private double deliveryVolume;
     private boolean fragile;
-    private double totalPrice;
-    private double deliveryPrice;
-    private double productPrice;
+    private BigDecimal totalPrice;
+    private BigDecimal deliveryPrice;
+    private BigDecimal productPrice;
 }
