+++ commerce/payment/src/main/java/ru/yandex/practicum/commerce/payment/controller/PaymentController.java
@@ -8,6 +8,7 @@ import ru.yandex.practicum.commerce.dto.PaymentDto;
 import ru.yandex.practicum.commerce.feign.PaymentClient;
 import ru.yandex.practicum.commerce.payment.service.PaymentService;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @RestController
@@ -18,12 +19,12 @@ public class PaymentController implements PaymentClient {
     private final PaymentService paymentService;
 
     @Override
-    public Double getTotalCost(OrderDto orderDto) {
+    public BigDecimal getTotalCost(OrderDto orderDto) {
         return paymentService.getTotalCost(orderDto);
     }
 
     @Override
-    public Double productCost(OrderDto orderDto) {
+    public BigDecimal productCost(OrderDto orderDto) {
         return paymentService.productCost(orderDto);
     }
